Moved middleware to helpers
The `check-sha` and `error-handler` middleware are now under `/lib/helpers`
